# Speed compilation by using .rmeta over .rlib files